backend/requirements.txt
# Avadhan Training Backend - Python/PyTorch

torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
onnxruntime-gpu>=1.15.0
safetensors>=0.4.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
pydantic>=2.0.0
python-multipart>=0.0.6
numpy>=1.24.0
faiss-cpu>=1.7.0
aiofiles>=23.0.0
tqdm>=4.65.0
accelerate>=0.20.0
bitsandbytes>=0.40.0
